+++ cmd/qm_edit.go
@@ -15,9 +15,10 @@ import (
 )
 
 var qmEditCmd = &cobra.Command{
-	Use:   "edit [name-or-vmid]",
-	Short: "Edit a VM's config in $EDITOR",
-	Args:  cobra.MaximumNArgs(1),
+	Use:               "edit [name-or-vmid]",
+	Short:             "Edit a VM's config in $EDITOR",
+	Args:              cobra.MaximumNArgs(1),
+	ValidArgsFunction: completeVMNames,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		client, err := loadClient()
 		if err != nil {